+++ gwcs/tags/tests/test_selector.py
@@ -6,7 +6,7 @@ from __future__ import absolute_import, division, unicode_literals, print_functi
 import numpy as np
 
 from astropy.modeling.models import Mapping, Shift, Scale, Polynomial2D
-from astropy.tests.helper import pytest
+import pytest
 from ... import selector, extension
 from asdf.tests import helpers
 from ...tests.test_region import create_range_mapper, create_scalar_mapper